2019-05-12 DB: modified code from Sun 12 May
