+++ App/Project.php
@@ -1,11 +1,13 @@
 <?php
-
 namespace App;
 
+use Exception;
 class Project
 {
     public const Table ="Project"; // <- value that can not be change
-
+	public const ERROR_PROP_GET = 'ERROR: tried to read a non-existent property: %s';
+	public const ERROR_PROP_SET = 'ERROR: tried to set a non-existent property: %s';
+	
     public $projectName; // <- Properties
     public $projectDateCreated; // <- Properties
     public $projectAuthor; // <- Properties
@@ -42,6 +44,19 @@ class Project
         return $this->projectAuthor;
     }
 
+	// protect against calls to non-existent properties
+	public function __get($prop)
+	{
+		$message = sprintf(self::ERROR_PROP_GET, $prop);
+		error_log($message);
+		throw new Exception($message);
+	}
+	public function __set($prop, $params)
+	{
+		$message = sprintf(self::ERROR_PROP_SET, $prop);
+		error_log($message);
+		throw new Exception($message);
+	}
 }
